+++ src/ecommerce/prices/precios/assets/Graficas.css
@@ -1,0 +1,49 @@
+.graficas {
+    overflow-x: auto;
+    margin-top: 30px;
+    background-color: hsl(0, 0%, 100%);
+    border: 1px solid hsl(240, 5.9%, 90%);
+    border-radius: 20px;
+    padding: 30px;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+.contenedor-graficas{
+    width: 40%;
+}
+.titulo-grafica {
+    font-size: 1.6rem;
+    margin-top: 0;
+    margin-bottom: 0;
+    font-weight: 600;
+}
+.info-grafica {
+    color: grey;
+}
+
+
+.btn-refrescar-graf {
+    background-color: #f0ad4e;
+    color: white; 
+    border: none;
+    border-radius: 5px; 
+    padding: 5px 10px; 
+    font-size: 15px; /* Tamaño del texto */
+    cursor: pointer; /* Cambiar a un puntero */
+    transition: background-color 0.3s ease; /* Efecto de transición */
+  }
+  
+
+  .btn-refrescar-graf:disabled {
+    background-color: #9e9e9e; /* Color gris para estado deshabilitado */
+    cursor: not-allowed; /* Cambiar el cursor */
+  }
+  
+
+  .info-graf-container {
+    display: flex; /* Flexbox para alinear los elementos horizontalmente */
+    align-items: center; /* Centrar los elementos verticalmente */
+    justify-content: space-between; /* Espaciado entre texto y botón */
+    gap: 10px; /* Espaciado entre los elementos */
+    
+  }
