+++ src/ecommerce/prices/precios/assets/Historial.css
@@ -1,0 +1,395 @@
+/* Contenedor principal */
+.historial {
+  font-family: Arial, sans-serif;
+  color: #333;
+  margin: 2em;
+  background-color:  hsl(0, 0%, 100%);
+  border-radius: 8px;
+  padding: 10px 30px 30px 30px;
+}
+
+/* Encabezado */
+.historial-header {
+  display: flex;
+  align-items: center;
+  padding: 1rem 1rem;
+  border-bottom: 1px solid #ddd;
+gap: 10px;
+}
+
+.historial-header i {
+ font-size: 1.5rem; 
+}
+.historial-title {
+  font-size: 1.6em;
+  font-weight: bold;
+  margin-right: auto;
+  margin: 10px auto 10px 10px;
+}
+
+.buscador {
+  padding: 0.5em;
+  width: 300px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+}
+.historial-header .EditarBtn{
+font-size: 15px;
+padding: 6px 20px 10px 20px;
+color: white;
+border-radius: 10px;
+font-weight: 700;
+background: #4f42a7;
+cursor: pointer;
+transition: transform 0.3s ease, color 0.3s ease;
+}
+
+.historial-header .EditarBtn i {
+  font-size: 1rem;
+  margin: 4px 6px 0px 0px;
+}
+.botones-acciones .boton-filtro {
+  background-color: #2196F3;
+}
+
+/* Tabla */
+.tabla-historial {
+  padding: 10px;
+  width: 100%;
+  border-collapse: collapse;
+}
+
+.tabla-historial th, .tabla-historial td {
+  padding: 1em;
+  border-bottom: 1px solid #ddd;
+  text-align: left;
+}
+
+.tabla-historial th {
+  background-color: #f5f5f5;
+  color: black;
+  font-weight: 600;
+}
+
+.fila-llamada:hover {
+  background-color: #eee;
+}
+
+/* Estilo para el tipo de cambio de modificación */
+.tipo-modificacion-agregado {
+  background-color: #c8f4ac;
+  font-weight: bold;
+  color: black;
+  padding: 5px 12px;
+  border-radius: 15px;
+  font-size: .8rem;
+}
+
+.tipo-modificacion-eliminado {
+  background-color: #ff7d7d;
+  font-weight: bold;
+  color: black;
+  padding: 5px 10px;
+  border-radius: 15px;
+  font-size: .8rem;
+}
+
+.tipo-modificacion-editado {
+  background-color: #ecff83;
+  font-weight: bold;
+  color: black;
+  padding: 5px 18px;
+  border-radius: 15px;
+  font-size: .8rem;
+}
+
+
+/* Estilo para la información del usuario */
+.info-usuario {
+  display: flex;
+  align-items: center;
+}
+
+.avatar-usuario {
+  background-color: #ddd;
+  border-radius: 50%;
+  width: 30px;
+  height: 30px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-right: 0.5em;
+  font-weight: bold;
+  color: #555;
+}
+
+/* Estilos para cada tipo de modificación */
+.modificacion-agregado {
+  background-color: green;
+}
+
+.modificacion-eliminado {
+  background-color: #fdecea;
+  color: red;
+}
+
+.modificacion-editado {
+  background-color: #fff3cd;
+  color: #856404;
+}
+
+/* Popup básico */
+.popup-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.5);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.popup-content {
+  background-color: white;
+  padding: 20px;
+  border-radius: 8px;
+  max-width: 600px;
+  width: 80%;
+}
+
+.boton-eliminar {
+  background-color: #e74c3c;
+  color: white;
+  border: none;
+  padding: 5px 10px;
+  cursor: pointer;
+  border-radius: 3px;
+}
+
+.boton-eliminar:hover {
+  background-color: #c0392b;
+}
+
+/* Estilos generales para el contenedor del formulario */
+form {
+background-color: #f9f9f9;
+border: 1px solid #ddd;
+border-radius: 8px;
+padding: 20px;
+max-width: 400px;
+margin: 20px auto;
+box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+/* Estilos para las etiquetas */
+form label {
+display: block;
+font-weight: bold;
+margin-bottom: 8px;
+color: #333;
+}
+
+/* Estilos para los inputs */
+form input {
+width: 100%;
+padding: 8px;
+margin-bottom: 16px;
+border: 1px solid #ccc;
+border-radius: 4px;
+font-size: 14px;
+}
+
+/* Estilos para los botones */
+form button {
+padding: 10px 15px;
+border: none;
+border-radius: 4px;
+font-size: 14px;
+cursor: pointer;
+margin-right: 10px;
+}
+
+/* Botón de guardar */
+form button[type="submit"] {
+background-color: #4CAF50;
+color: white;
+}
+
+/* Botón de cancelar */
+form button[type="button"] {
+background-color: #f44336;
+color: white;
+}
+
+/* Hover effects para los botones */
+form button[type="submit"]:hover {
+background-color: #45a049;
+}
+
+form button[type="button"]:hover {
+background-color: #d32f2f;
+}
+
+.historial .popup-content {
+max-width: 800px;
+max-height: 90vh;
+overflow-y: auto;
+background-color: #fff;
+padding: 20px;
+border-radius: 8px;
+box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+}
+
+.modal-form {
+margin-top: 10px;
+}
+
+.modal-form form {
+display: flex;
+flex-direction: column;
+gap: 10px;
+}
+
+.modal-form input {
+padding: 8px;
+border: 1px solid #ccc;
+border-radius: 4px;
+}
+
+button {
+cursor: pointer;
+}
+
+button[type="submit"] {
+background-color: #4caf50;
+color: white;
+}
+
+button[type="button"] {
+background-color: #f44336;
+color: white;
+}
+
+.boton-eliminar {
+background-color: #e74c3c;
+color: white;
+border: none;
+padding: 5px 10px;
+border-radius: 4px;
+}
+
+button.editar-btn {
+background-color: #3498db;
+color: white;
+border: none;
+padding: 5px 10px;
+border-radius: 4px;
+}
+button.CERRAR {
+background-color: #e74c3c; /* Rojo llamativo para atraer atención */
+color: white; /* Texto blanco para contraste */
+border: none; /* Sin borde */
+padding: 10px 20px; /* Tamaño del botón */
+font-size: 16px; /* Tamaño de la fuente */
+border-radius: 5px; /* Bordes redondeados */
+cursor: pointer; /* Cambiar cursor al pasar por encima */
+transition: background-color 0.3s ease; /* Animación de color */
+}
+
+button.CERRAR:hover {
+background-color: #c0392b; /* Más oscuro al pasar el mouse */
+}
+
+button.CERRAR:active {
+background-color: #a93226; /* Aún más oscuro al hacer clic */
+}
+
+button.CERRAR:focus {
+outline: none; /* Sin borde al enfocarse */
+box-shadow: 0 0 5px 2px rgba(231, 76, 60, 0.5); /* Efecto de enfoque */
+}
+/* Estilo general para la tabla */
+.popup-content table {
+width: 100%; /* Usar todo el ancho disponible */
+border-collapse: collapse; /* Unir bordes de celdas */
+margin-top: 10px; /* Separación superior */
+font-family: Arial, sans-serif;
+font-size: 14px;
+}
+
+.popup-content th, 
+.popup-content td {
+border: 1px solid #ddd; /* Bordes suaves */
+padding: 8px; /* Espaciado dentro de las celdas */
+text-align: left; /* Alinear texto a la izquierda */
+}
+
+.popup-content th {
+background-color: #f4f4f4; /* Color de fondo para encabezados */
+color: #333; /* Color del texto de los encabezados */
+font-weight: bold; /* Texto en negrita */
+}
+
+.popup-content tr:nth-child(even) {
+background-color: #f9f9f9; /* Fondo diferente para filas pares */
+}
+
+.popup-content tr:hover {
+background-color: #f1f1f1; /* Fondo al pasar el ratón */
+}
+
+/* Estilo para botones dentro de la tabla */
+.popup-content .boton-eliminar {
+background-color: #e74c3c; /* Rojo llamativo */
+color: white;
+border: none;
+padding: 5px 10px;
+border-radius: 3px;
+cursor: pointer;
+font-size: 12px;
+}
+
+.popup-content .editar-btn {
+background-color: #3498db; /* Azul para editar */
+color: white;
+border: none;
+padding: 5px 10px;
+border-radius: 3px;
+cursor: pointer;
+font-size: 12px;
+}
+
+.popup-content .boton-eliminar:hover {
+background-color: #c0392b; /* Más oscuro al pasar el mouse */
+}
+
+.popup-content .editar-btn:hover {
+background-color: #2980b9; /* Más oscuro al pasar el mouse */
+}
+
+/* Estilo para el contenedor de la modal */
+.popup-overlay {
+position: fixed;
+top: 0;
+left: 0;
+width: 100%;
+height: 100%;
+background: rgba(0, 0, 0, 0.5); /* Fondo semitransparente */
+display: flex;
+align-items: center;
+justify-content: center;
+z-index: 1000;
+}
+
+.popup-content {
+background: white;
+padding: 20px;
+border-radius: 10px;
+max-width: 800px;
+width: 90%;
+box-shadow: 0 0 15px rgba(0, 0, 0, 0.3);
+overflow-y: auto;
+max-height: 90%; /* Limitar altura */
+}
