+++ src/ecommerce/prices/precios/assets/routesTabs.css
@@ -1,0 +1,135 @@
+
+
+.navbarRoutes {
+    margin: 10px 10px 0px 10px;
+    display: flex;
+    flex-direction: column;
+    background-color: #fff;
+    padding: 10px 20px;
+    border-radius: 8px;
+    align-items: center;
+  
+}
+
+.navbarRoutes-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+  
+}
+
+.navbarRoutes-hamburger {
+    background: none;
+    border: none;
+    font-size: 24px;
+    cursor: pointer;
+    display: none; /* Oculto por defecto */
+}
+
+.navbarRoutes-links {
+    display: flex;
+    width: 100%;
+    gap: 10px;
+    justify-content: space-around;
+    flex-wrap: wrap;
+}
+
+.navbarRoutes-item {
+    color: #000;
+    text-decoration: none;
+    padding: 10px 15px;
+    display: flex;
+    align-items: center;
+    font-size: 1rem;
+    transition: color 0.3s ease;
+}
+
+.navbarRoutes-item i {
+    margin-right: 5px;
+}
+
+.navbarRoutes-item:hover {
+    color: #4A4AFF;
+}
+
+.navbarRoutes-item.active {
+    color: #4A4AFF;
+    position: relative;
+}
+
+.navbarRoutes-item.active::after {
+    content: '';
+    width: 100%;
+    height: 3px;
+    background-color: #4A4AFF;
+    position: absolute;
+    bottom: -5px;
+    left: 0;
+    border-radius: 5px;
+}
+
+
+
+.nav-logo {
+    display: flex;
+    align-items: center;
+    gap: 6px;
+
+  }
+  
+  .nav-imagen {
+    width: 30px;
+    height: 30px;
+}
+
+  
+  .nav-titulo {
+    font-weight: 500;
+    font-size: .8rem; 
+  }
+
+  
+  .nav-usuario {
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+  }
+  
+  .nav-foto {
+    border-radius: 50%; 
+    width: 40px;
+    height: 40px;
+  }
+  
+  .nav-nombre {
+    font-weight: 700;
+    font-size: 17px;
+  }
+
+  .nav-usuario2{
+    display: none;
+}
+
+  
+
+
+/* Estilos responsivos */
+@media (max-width: 768px) {
+    .navbarRoutes-links {
+        display: none; /* Oculto en pantallas pequeñas */
+        flex-direction: column;
+        gap: 0;
+    }
+    .navbarRoutes-links.open {
+        display: flex; /* Mostrado cuando se abre el menú */
+    }
+    .navbarRoutes-hamburger {
+        display: block; /* Mostrar el botón de hamburguesa en pantallas pequeñas */
+        color: #000;
+    }
+    .navbarRoutes {
+        flex-direction: column;
+        align-items: flex-start;
+    }
+}
